refactor:  rename ui folders
