+++ projects/movies/src/app/app-shell/app-shell.component.scss
@@ -1,6 +1,6 @@
-@import '../ui/tokens/mixins';
-@import '../ui/tokens/flex';
-@import '../ui/atoms/icons/icon';
+@import '../ui/token/mixins';
+@import '../ui/token/flex';
+@import '../ui/component/icons/icon';
 
 
 .app-toolbar {
