+++ projects/movies/src/app/pages/movie-detail-page/movie-detail-page.component.scss
@@ -1,6 +1,6 @@
-@import '../../ui/atoms/button/button';
-@import '../../ui/atoms/icons/icon';
-@import "../../ui/tokens/flex";
+@import '../../ui/component/button/button';
+@import '../../ui/component/icons/icon';
+@import "../../ui/token/flex";
 
 :host {
   width: 100%;
