+++ projects/movies/src/app/pages/movie-list-page/movie-list-page.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MovieListPageComponent } from './movie-list-page.component';
 import { RouterModule, Routes } from '@angular/router';
-import { MovieListModule } from '../../ui/components/movie-list/movie-list.module';
+import { MovieListModule } from '../../ui/pattern/movie-list/movie-list.module';
 import { LetModule } from '@rx-angular/template/let';
-import { LoaderComponentModule } from '../../ui/atoms/loader/loader.component';
+import { LoaderComponentModule } from '../../ui/component/loader/loader.component';
 
 const ROUTES: Routes = [
   {
