--- projects/movies/src/app/ui/atoms/aspect-ratio-box/aspect-ratio-box.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { AspectRatioBoxComponent } from './aspect-ratio-box.component';
-
-@NgModule({
-  declarations: [AspectRatioBoxComponent],
-  imports: [CommonModule],
-  exports: [AspectRatioBoxComponent],
-})
-export class AspectRatioBoxModule {}
