--- projects/movies/src/app/ui/atoms/backdrop/backdrop.component.scss
@@ -1,21 +1,0 @@
-@import "../../tokens/zIndex";
-@import "../../tokens/flex";
-
-:host {
-  visibility: hidden;
-  opacity: 0;
-  position: fixed;
-  z-index: var(--theme-backdrop-zIndex);
-  @include d-flex-vh;
-  right: 0;
-  bottom: 0;
-  top: 0;
-  left: 0;
-  background-color: var(--palette-background-backdrop);
-  transition: opacity var(--theme-anim-duration-leavingScreen) var(--theme-anim-easing-easeInOut) 0ms;
-}
-
-:host.opened {
-  visibility: visible;
-  opacity: 1;
-}
