--- projects/movies/src/app/ui/atoms/button/_button.scss
@@ -1,22 +1,0 @@
-@import "../../tokens/flex";
-.btn {
-  @include d-flex-vh;
-  @include d-inline-flex;
-  outline: none;
-  padding: 6px 16px;
-  min-width: 96px;
-  min-height: 48px;
-  font-weight: normal;
-  font-size: var(--text-md);
-  color: var(--palette-primary-main);
-  border: 1px solid rgba(var(--palette-primary-main-rgb), 0.5);
-  border-radius: 4px;
-  box-shadow: none;
-  background-color: transparent;
-}
-
-.btn__icon {
-  width: 24px;
-  height: 24px;
-  margin-left: 8px;
-}
