--- projects/movies/src/app/ui/atoms/dark-mode-toggle/dark-mode-toggle.module.ts
@@ -1,11 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { DarkModeToggleComponent } from './dark-mode-toggle.component';
-import { LetModule } from '@rx-angular/template/let';
-
-@NgModule({
-  declarations: [DarkModeToggleComponent],
-  exports: [DarkModeToggleComponent],
-  imports: [CommonModule, LetModule],
-})
-export class DarkModeToggleModule {}
