--- projects/movies/src/app/ui/atoms/detail-grid/detail-grid.component.ts
@@ -1,18 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-
-@Component({
-  selector: 'ds-detail-grid',
-  template: `
-    <div class="grid--item gradient">
-      <ng-content select="[detailGridMedia]"></ng-content>
-    </div>
-    <div class="grid--item media">
-      <ng-content select="[detailGridDescription]"></ng-content>
-    </div>
-  `,
-  styleUrls: ['./detail-grid.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class DetailGridComponent {
-
-}
