--- projects/movies/src/app/ui/atoms/detail-grid/detail-grid.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { GridComponent } from './subtitle.component';
-
-@NgModule({
-  declarations: [GridComponent],
-  imports: [],
-  exports: [GridComponent]
-})
-export class DetailGridModule {
-}
