--- projects/movies/src/app/ui/atoms/form-control/form-comtrol.component.scss
@@ -1,10 +1,0 @@
-@import "../../tokens/flex";
-:host {
-  margin: 8px 0;
-  @include d-flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-:host > :global(label + *) {
-  margin-top: 4px;
-}
