--- projects/movies/src/app/ui/atoms/form-control/form-control.component.ts
@@ -1,15 +1,0 @@
-import { ChangeDetectionStrategy, Component, NgModule } from '@angular/core';
-
-@Component({
-  selector: 'app-form-control',
-  template: `<ng-content></ng-content>`,
-  styles: ['./form-control.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class FormControlComponent {}
-
-@NgModule({
-  declarations: [FormControlComponent],
-  exports: [FormControlComponent],
-})
-export class FormControlComponentModule {}
