--- projects/movies/src/app/ui/atoms/hamburger-button/hamburger-button.component.html
@@ -1,3 +1,0 @@
-<span class="bar"></span>
-<span class="bar"></span>
-<span class="bar"></span>
