--- projects/movies/src/app/ui/atoms/hamburger-button/hamburger-button.component.scss
@@ -1,17 +1,0 @@
-@import "../../tokens/flex";
-
-:host {
-  @include d-flex;
-  flex-direction: column;
-  align-self: center;
-  justify-content: space-around;
-  width: 25px;
-  cursor: pointer;
-}
-.bar {
-  width: 100%;
-  height: 4px;
-  margin: 2px 0;
-  border-radius: 4px;
-  background-color: var(--palette-secondary-main);
-}
