--- projects/movies/src/app/ui/atoms/icons/_icon.scss
@@ -1,5 +1,0 @@
-.icon {
-  width: 24px;
-  height: 24px;
-  margin-right: 0.5rem;
-}
