--- projects/movies/src/app/ui/atoms/side-drawer/side-drawer.component.scss
@@ -1,54 +1,0 @@
-@import '../../tokens/mixins';
-@import "../../tokens/flex";
-
-:host {
-  display: contents;
-}
-.side-drawer {
-  position: fixed;
-  @include d-flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 250px;
-  max-width: 70%;
-  height: 100%;
-  left: 0;
-  top: 0;
-  z-index: var(--theme-drawer-zIndex);
-  padding: 32px 16px 16px;
-  overflow-y: auto;
-  box-sizing: border-box;
-  transition: transform var(--theme-anim-duration-short)
-    var(--theme-anim-easing-easeOut);
-  box-shadow: var(--theme-shadow-16);
-  background-color: var(--palette-background-paper);
-  contain: strict;
-  transform: translateX(-100%);
-
-  & + footer{
-    transition: transform var(--theme-anim-duration-short)
-      var(--theme-anim-easing-easeOut);
-    transform: translateX(-100%);
-  }
-
-  @include isDesktop {
-    box-shadow: unset;
-    background-color: transparent;
-  }
-}
-
-@media screen and (min-width: 1298px) {
-  .side-drawer {
-    transform: translateX(0);
-  }
-  app-backdrop {
-    display: none;
-  }
-}
-.side-drawer.opened {
-  transform: translateX(0);
-
-  & + footer{
-    transform: translateX(0);
-  }
-}
