--- projects/movies/src/app/ui/atoms/star-rating/star-rating.component.scss
@@ -1,62 +1,0 @@
-@import './tooltip';
-@import "../../tokens/flex";
-@import "../../tokens/colors";
-
-:host {
-  position: relative;
-  @include d-flex-vh;
-
-  &:hover .tooltip {
-    visibility: visible;
-  }
-}
-
-.tooltip {
-  top: -28px;
-}
-
-.rating-value {
-  overflow: hidden;
-  color: var(--palette-warning-main);
-  font-size: var(--text-sm);
-  padding: 5px 0 0 5px;
-}
-
-.stars {
-  overflow: hidden;
-  position: relative;
-}
-
-.star {
-  position: relative;
-  overflow: hidden;
-  cursor: pointer;
-  display: block;
-  float: left;
-  font-size: 24px;
-  color: var(--palette-warning-main);
-}
-
-.star-full {
-  color: var(--palette-warning-main);
-}
-
-.star-half {
-  color: var(--palette-text-secondary);
-
-  &:before {
-    position: absolute;
-    overflow: hidden;
-    display: block;
-    z-index: 1;
-    top: 0;
-    left: 0;
-    width: 50%;
-    content: '★';
-    color: var(--palette-warning-main);
-  }
-}
-
-.star-empty {
-  color: var(--palette-text-secondary);
-}
