--- projects/movies/src/app/ui/atoms/star-rating/star-rating.module.ts
@@ -1,11 +1,0 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {StarRatingComponent} from './star-rating.component';
-
-@NgModule({
-  declarations: [StarRatingComponent],
-  imports: [CommonModule],
-  exports: [StarRatingComponent],
-})
-export class StarRatingModule {
-}
