--- projects/movies/src/app/ui/components/movie-list/movie-list.component.scss
@@ -1,76 +1,0 @@
-@import "../../tokens/flex";
-
-.movies-list--grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(10rem, 25rem));
-  gap: 4rem 2rem;
-  place-content: space-between space-evenly;
-  align-items: start;
-  margin: 4rem 0;
-  position: relative;
-  contain: layout;
-
-  @media only screen and (max-width: 600px) {
-    grid-template-columns: repeat(auto-fit, minmax(10rem, 23rem));
-    justify-content: space-around;
-    gap: 4rem 1.5rem;
-  }
-
-  @media only screen and (max-width: 500px) {
-    grid-template-columns: repeat(auto-fit, minmax(10rem, 18rem));
-    gap: 4rem 1rem;
-  }
-
-  &-item {
-    @include d-flex;
-    align-items: stretch;
-    flex-direction: column;
-    transition: transform 150ms cubic-bezier(0.4, 0, 0.2, 1) 0s;
-
-    &:after {
-      content: '';
-      position: absolute;
-      z-index: -99;
-      top: 0px;
-      left: 0px;
-      width: 100%;
-      height: 100%;
-      transform: scaleY(0);
-      transform-origin: center top;
-      background-color: var(--palette-background-paper);
-      box-shadow: rgb(0 0 0 / 20%) 0 2px 1px -1px,
-        rgb(0 0 0 / 14%) 0px 1px 1px 0px, rgb(0 0 0 / 12%) 0px 1px 3px 0px;
-    }
-
-    &-image {
-      position: relative;
-      img {
-        width: 100%;
-        height: auto;
-      }
-    }
-
-    &__details {
-      @include d-flex-v;
-      flex-direction: column;
-      justify-content: space-between;
-      padding: 1.5rem 3rem;
-      @media only screen and (max-width: 500px) {
-        padding: 1.5rem;
-      }
-
-      &-title {
-        text-align: center;
-        margin-bottom: 1rem;
-        line-height: 1.4;
-      }
-    }
-
-    &:hover {
-      transform: scale(1.03);
-      &:after {
-        transform: scale(1);
-      }
-    }
-  }
-}
