+++ projects/movies/src/app/ui/token/_animations.scss
